builtin.c: Fix C++ compilation errors
C++ does not allow `goto` across initalisations of variables that are
in scope at the label, so just get rid of the `goto`.

An alternative would be to start a new scope after the first `goto`,
so that the variables aren't in scope any more at the label, but I
think this is neater.