Test equivalence of 'true' and 'false' to negated statements
The documentation for lib/builtin.pm asserts that 'true' is equivalent
to !!1 or !0, and that false is equivalent to !!0 or !1.  Demonstrate
that this is so.

Per review by Paul Evans in https://github.com/Perl/perl5/pull/19252,
we'll use cmp_ok() rather than is() and test for each of 'eq' and '=='.