Change pack U behavior for EBCDIC
This effectively reverts 3ece276e6c0.

It turns out this was a bad idea to make U mean the non-native official
Unicode code points.  It may seem to make sense to do so, but broke
multiple CPAN modules which were using U the previous way.

This commit has no effect on ASCII-platform functioning.