make a sort with zero args a compile-time err
    @a = sort @empty; # unaffected
    @a = sort;        # now a compile-time error

the main motivation for this is to potentially be able to treat a
colon following the 'sort' keyword as part of the sort expression,
to then be able to add sort attributes at some point, e.g.
    @a = sort :num, 1,2,3;
rather than the colon being seen as part of a ?: conditional.

See http://nntp.perl.org/group/perl.perl5.porters/262425.