Tests for Issue #19350, assert fail with Debug=>ALL
Note that 'use re Debug => "ALL"' and 'use re Debug => "PARSE"'
are basically the same as far as this issue goes. ALL implies PARSE
along with other things, the fault is in how the PARSE part works,
and the omitting the full output of ALL and just testing PARSE is
sufficient to test this issue, and avoids any need to deal with
memory addresses or word size issues that are irrelevant to the
issue anyway.