sv.c: newSV* functions now use sv_upgrade_fresh
Most of the various newSV* functions modified in this commit all aquire
a new SV head and upgrade it to a specific PV type of choice. Changing
the call from sv_upgrade(sv, type) to sv_upgrade_fresh(sv,type) allows
for the inlining of only the relevant body allocation code into these
functions.

The exception to the above is Perl_newSV_type, which takes the target
type as a parameter. For some callers, the target types may not be known
until runtime. In this case, the entirety of Perl_sv_upgrade_fresh will
be inlined.

The growth in the size of the compiled perl binary is small (only about
40 bytes, at least on x64 using gcc or clang).