Add extra paired string delimiters feature
When this feature is enabled, one can use many more string delimiters
that have an opening version and a mirrored closing one.

The code in toke.c had two closely related loops; one for unmirrored
delimiters (same on both ends of the string) that could take UTF-8
delimiters, and one that allowed a mirrored closing delimiter, which
could take only a single byte.  I found that it was just easiest to
collapse these into one loop.