toke.c: Rmv unnecessary SV
This code created an SV simply to get SVfARG() to print it out.  (SVfARG
seems buggy, and I will look at it and fix or file a ticket if so.)  But
nowadays there is UTF8fARG which does the right thing on strings, so the
SV is unnecessary.

The code also has a fix where non-ASCII but Latin1 delimiters  when 'use
utf8' is in effect would potentially come out as garbage  This commit
also adds a clause in a conditional to prevent that.