regcomp.c: Properly optimize verb style lookahead/lookbehind
This introduces the full set of optimizations for verb style
lookahead and lookbehind:

Eg. (*nla:) and (*negative_lookahead:)  are equivalent to OPFAIL,
and (*plb:) and (*positive_lookbehind:) are equivalent to NOTHING,
and (*nlb:) and (*negative_lookbehind:) are equivalent to OPFAIL,
and (*pla:) and (*positive_lookahead:)  are equivalent to NOTHING.

At the same time it fixes a very minor bug that RExC_seen_zerowidth
and RExC_in_lookaround should not be incremented when
(*negative_lookahead:) is optimized to OPFAIL.

This patch makes the flag setting and increments consistent with the
logic for (?<=) and (?<!) and (?=) and (?!).