regex_sets: Allow all legal flags in nested calls
See https://github.com/Perl/perl5/issues/16779#issuecomment-563219424

This commit extends the flags accepted by nested calls to regex sets to
any legal set.  Previously it allowed only '^', as that is what an
actual compilation would return.  But as pointed out in the conversation
in that ticket, it's possible to detach flags from the rest of the
pattern, and to write equivalent ones in multiple ways.

This commit just changes one line to look for any legal flag, not
restricting it to '^'.