regcomp.c: rename "fake" where it used in recursive study_chunk() calls
All instances of the fake variable are used inside of data_fake.last_closep
structures, so rename them from "fake" to "fake_last_close" so it is obvious
what they are used for. Note the 'p' in last_closep means "pointer", eg
it means "pointer to last_close", so the "fake" version is called
"fake_last_close" as well.