test.pl - allow more than 676 tempfiles() in one process
This reworks testfile() to allow more than 26**2 (676) calls to
tempfile() in the course of a single test process.

In order to keep the filenames short (which I believe is an objective of
this code) I switched the presentation of the pid to using base 26 as
well. Also added comments and a nicer error message if you reach the new
max of 26**3 (17576)