hv_func.h - switch sbox32 initializer to use the 128 bit seeding logic
The 96 bit seed initializer was not mixing the seed bits very well and a
a 1 bit change to the seed used for the sbox initialization was not
propagating throughout the full SBOX buffer. The 128 bit variant is more
robust, and this patch makes it even more so by churning it longer, and
mixing in the seed twice.

This change also removes the old 96 bit variant and code that uses it
as I don't think it should be used anymore at all.