pod/perldata.pod - document demarcated var form (curly braces) better
Perl supports using curly braces for demarcated variable notation, both
in normal code but especially for interpolation. Some variables, such
as a multi-character caret variables must use this form at all times,
and must use this form to achieve interpolation of a variable subscript.

Prior to this patch we only documented the simple forms of this feature,
for example "${who}'s this then", but we did not document that it also
includes more complex variable access like array subscripts
"@{array[1,5]}" nor the cases where this form is outright required such
as when interpolating multi char caret variables like "${^CAPTURE[10]}".

This patch provides explanation for the full range of cases and also
documents how backslash can used as another form of disambiguation.

See GH Issue #19497 and GH Issue #18800 for examples of people being
confused by this notation.