t/op/magic.t - Issue #19504 - fixup tests for utf8 $0
Since PR #19334 was merged the tests for it are failing. This should fix
them. I am taking the perspective based on the broken previous tests
that that on some systems /proc/$$/commandline returns a trailing null.
This fixes the code to read /proc/$$/comandline to chop the null off if
it is there so the tests can be agnostic to that.

At the same time I have fixed the tests a bit to not use regexen where
they aren't needed and to avoid a test count issue in the skip logic,
and also to produce better errors if the code to read
/proc/$$/commandline fails.