manicheck: fix broken exitcode logic
The comment says the highest exit code we should return is 124, but the
code will return 125 if there are 125 problems, if it is higher than 125
it will return 124. Which doesn't make sense.

This patch changes the logic to do what the comment says, return 124 if
there are 124 or more problems, otherwise it returns the count of
problems.