Add documentation on slices on an array or hash element
In PR #18827 Dan Jacobson requested we provide examples for slices on an
array element or hash element, he struggled to get the patch right so I
am pushing this as a replacement.

The feedback to PR #18827 suggested including a brief introduction to
postfix notation so I added that as well as a reference to the
perlref manpage.