regcomp.c: reorder and comment S_study_chunk() internal vars
The list was kinda random and did not include many comments, which made
life difficult understanding what the purpose of the different vars.

This documents them, and includes a follow up to investigate that came
up during the documentation process: first_non_open seems a bit off.
Will investigate and improve further in a subsequent patch.