regexec.c: Fix up for ACCEPT inside of a (...)+ set lastopen in CURLYM
GH Issue #19484 reported that

    print "ABDE" =~ /(A (A|B(*ACCEPT)|C)+ D)(E)/x ? "yes: <$1-$2>" : "no";

does not output the expected 'AB-B', and instead does not match.
Removing the + quantifier behaves as expected.

This patch is 2/4 and fixes part of the problem: lastopen was not
being set properly inside of the CURLYM optimization. lastopen is used
by the ACCEPT logic to know which parens need to be closed.